+++ src/utils.ts
@@ -528,7 +528,7 @@ export async function getUsernameInstanceUrl(
     {
       fail: false,
       output: false,
-      cacheExpiration: 1000 * 60 * 60 * 24, // 1 day
+      cacheExpiration: 1000 * 60 * 60 * 24, // 1 day (milliseconds)
       cacheSection: "project",
     },
   );