+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## [6.25.2] 2025-01-26
+
+- Deactivate and re-active sf-git-merge-driver during plugin upgrade.
+- Pipeline view: Use `sf org open` command to open orgs when clicking in the diagram, to benefit from existing user authentication.
+
 ## [6.25.1] 2025-01-21
 
 - Refactor command groups for improved navigation in context menus